rename SpuController to SpuHealthCheckController
