+++ crates/fluvio-sc/src/controllers/health_check/mod.rs
@@ -1,5 +1,3 @@
-mod actions;
 mod controller;
 
-pub use self::actions::*;
 pub use self::controller::*;
