+++ crates/fluvio-sc/src/controllers/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod partitions;
-pub(crate) mod spus;
+pub(crate) mod health_check;
 pub(crate) mod topics;
 pub(crate) mod scheduler;
